kafka_processing/requirements.txt
Flask
kafka-python
paramiko
pandas
python-dotenv
psycopg2-binary
streamlit
scikit-learn
joblib
